+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Added Ring Loss override in the mod settings.
 - Added Custom songs only as a source for music randomization.
 - Added support for 1.1.1 seeds using overrides.
+- Any non-mapped ADX file from the custom folder will be added to the pool as a Level type song.
 
 ### Fixed: